Add files via upload
This project shows the top 10 countries with the highest Covid new cases and deaths between December 2019 and June 2020. The Covid data consist of different countries, but I decided to focus on the top 10 countries for this project. 

The outcome of this project is a dashboard showing the top 10 countries with new cases and death rates, the number of aged 65 and above per cent, the number of hospital beds available, GDP per capita for the 10 countries and also the total increased rate of new cases and deaths from December 2019 to June 2020. It also shows the total world's new cases and deaths.